+++ packages/common/src/deserializeSchema.ts
@@ -1,5 +1,5 @@
 import { Readable } from 'stream';
-import { ParameterType } from './types';
+import { ParameterType, SchemaVersion } from './types';
 import { PassThrough } from 'stream';
 import { Buffer } from 'buffer/';
 /**
@@ -12,27 +12,60 @@ export interface Deserial<T> {
 }
 
 /**
- * Map of contract names to contract schemas.
+ * schema (V1) for a module, which is a
+ * map of contract names to contract schemas.
  */
-export type Module = Record<string, Contract>;
+export type ModuleV1 = Record<string, ContractV1>;
+/**
+ * schema (V2) for a module, which is a
+ * map of contract names to contract schemas.
+ */
+export type ModuleV2 = Record<string, ContractV2>;
 
 /**
- * Reads {@link Module} from the given {@link Readable}.
+ * Versioned schema for a module.
+ * Note that the version is the schema version and not the module version.
+ */
+export type VersionedModule =
+    | {
+          v: SchemaVersion.V1;
+          value: ModuleV1;
+      }
+    | {
+          v: SchemaVersion.V2;
+          value: ModuleV2;
+      };
+
+/**
+ * Reads the schema (V1) for a contract module from the given {@link Readable}.
  *
  * @param source input stream
- * @returns module (contract map)
+ * @returns schema (V1) of a module (contract map)
  */
-export function deserialModule(source: Readable): Module {
-    return deserialMapFn<string, Contract>(
+export function deserialModuleV1(source: Readable): ModuleV1 {
+    return deserialMapFn<string, ContractV1>(
         deserialString,
-        deserialContract
+        deserialContractV1
     )(source);
 }
 
 /**
- * Contract schema.
+ *  Reads the schema (V2) for a contract module from the given{@link Readable}.
+ *
+ * @param source input stream
+ * @returns schema (V2) of a module (contract map)
  */
-export type Contract = {
+export function deserialModuleV2(source: Readable): ModuleV2 {
+    return deserialMapFn<string, ContractV2>(
+        deserialString,
+        deserialContractV2
+    )(source);
+}
+
+/**
+ *   schema (V1) for a contract.
+ */
+export type ContractV1 = {
     /** Optional schema for the contract state. */
     state: Type | null;
     /** Optional schema for init function parameters. */
@@ -42,18 +75,44 @@ export type Contract = {
 };
 
 /**
- * Reads {@link Contract} from the given {@link Readable}.
+ *  schema (V2) for a contract.
+ */
+export type ContractV2 = {
+    /** Optional schema for init function. */
+    init: ContractFunction | null;
+    /** Map of receive function names to schemas for their respective parameters and return values. */
+    receive: Record<string, ContractFunction>;
+};
+
+/**
+ * Reads {@link ContractV1} from the given {@link Readable}.
  *
  * @param source input stream
- * @returns contract
+ * @returns schema (V1) for a contract.
  */
-export function deserialContract(source: Readable): Contract {
+export function deserialContractV1(source: Readable): ContractV1 {
     return {
         state: deserialOptionFn<Type>(deserialType)(source),
         init: deserialOptionFn<Type>(deserialType)(source),
         receive: deserialMapFn<string, Type>(
             deserialString,
             deserialType
+        )(source),
+    };
+}
+
+/**
+ * Reads {@link ContractV2} from the given {@link Readable}.
+ *
+ * @param source input stream
+ * @returns schema (V2) for a contract.
+ */
+export function deserialContractV2(source: Readable): ContractV2 {
+    return {
+        init: deserialOptionFn<ContractFunction>(deserialFunction)(source),
+        receive: deserialMapFn<string, ContractFunction>(
+            deserialString,
+            deserialFunction
         )(source),
     };
 }
@@ -103,6 +162,11 @@ export type Type =
     | StructType
     | EnumType
     | StringType;
+
+export type ContractFunction = {
+    parameter?: Type;
+    returnValue?: Type;
+};
 
 export type StringType = {
     typeTag:
@@ -163,6 +227,42 @@ export type PairType = {
     ofLeft: Type;
     ofRight: Type;
 };
+
+/**
+ * The first byte of the function denotes whether the parameters, return value, or both are typed.
+ */
+enum FunctionTag {
+    Parameter = 0,
+    ReturnValue = 1,
+    Both = 2,
+}
+
+/**
+ * Reads {@link ContractFunction} from the given {@link Readable}.
+ *
+ * @param source input stream
+ * @returns Function schema type
+ */
+export function deserialFunction(source: Readable): ContractFunction {
+    const tag: FunctionTag = deserialUint8(source);
+    switch (tag) {
+        case FunctionTag.Parameter:
+            return {
+                parameter: deserialType(source),
+            };
+        case FunctionTag.ReturnValue:
+            return {
+                returnValue: deserialType(source),
+            };
+        case FunctionTag.Both:
+            return {
+                parameter: deserialType(source),
+                returnValue: deserialType(source),
+            };
+        default:
+            throw new Error('Incorrect tag for function');
+    }
+}
 
 /**
  * Reads {@link Type} from the given {@link Readable}.
@@ -463,12 +563,45 @@ export function deserialUint32(source: Readable): number {
 }
 
 /**
- *
  * @param buffer wasm file buffer
+ * @param moduleVersion the version of the module
  * @returns deserialized module of wasm file
  */
-export function deserialModuleFromBuffer(buffer: Buffer): Module {
+export function deserialModuleFromBuffer(
+    buffer: Buffer,
+    schemaVersion: SchemaVersion
+): VersionedModule {
     const bufferStream = new PassThrough();
     bufferStream.end(buffer);
-    return deserialModule(bufferStream);
+    switch (schemaVersion) {
+        case SchemaVersion.V1:
+            return {
+                v: SchemaVersion.V1,
+                value: deserialModuleV1(bufferStream),
+            };
+        case SchemaVersion.V2:
+            return {
+                v: SchemaVersion.V2,
+                value: deserialModuleV2(bufferStream),
+            };
+        default:
+            throw new Error('Unsupported module version');
+    }
+}
+
+export function getParameterType(
+    schema: ContractFunction | Type | null,
+    schemaVersion: SchemaVersion
+): Type | null {
+    if (!schema) {
+        return null;
+    }
+    switch (schemaVersion) {
+        case SchemaVersion.V1:
+            return schema as Type;
+        case SchemaVersion.V2:
+            return (schema as ContractFunction).parameter || null;
+        default:
+            throw new Error('Unsupported module version');
+    }
 }