+++ packages/common/src/types.ts
@@ -1469,3 +1469,8 @@ export interface IdentityInput {
     idCredSecret: string;
     randomness: string;
 }
+
+export enum SchemaVersion {
+    V1 = 0, // Used by version 0 smart contracts.
+    V2 = 1, // Used by version 1 smart contracts.
+}