+++ packages/nodejs/test/smart-contract-tests/Parameter_All.test.ts
@@ -121,13 +121,14 @@ test('Parameter of data types with the wrong private key', async () => {
         },
         s_string: 'some string',
     };
-    const modulefileBuffer = getModuleBuffer(
+    const moduleFileBuffer = getModuleBuffer(
         'test/resources/schemaFiles/schema_test.bin'
     );
     const inputParams = serializeInitContractParameters(
         contractName,
         userInput,
-        modulefileBuffer
+        moduleFileBuffer,
+        0
     );
     const baseEnergy = 300000n;