+++ packages/nodejs/test/smart-contract-tests/Parameter_ContractAddress.test.ts
@@ -40,13 +40,14 @@ test('Parameter of ContractAddress with the wrong private key', async () => {
 
     const userInput = contractAddress;
 
-    const modulefileBuffer = getModuleBuffer(
+    const moduleFileBuffer = getModuleBuffer(
         'test/resources/schemaFiles/schema16.bin'
     );
     const inputParams = serializeInitContractParameters(
         contractName,
         userInput,
-        modulefileBuffer
+        moduleFileBuffer,
+        0
     );
     const baseEnergy = 300000n;