+++ packages/nodejs/test/smart-contract-tests/Parameter_Struct_U8_3Variables.test.ts
@@ -42,13 +42,14 @@ test('Parameter of Struct (3 varaibles) with the wrong private key', async () =>
         physics: 23,
     };
 
-    const modulefileBuffer = getModuleBuffer(
+    const moduleFileBuffer = getModuleBuffer(
         'test/resources/schemaFiles/schema12.bin'
     );
     const inputParams = serializeInitContractParameters(
         contractName,
         userInput,
-        modulefileBuffer
+        moduleFileBuffer,
+        0
     );
     const baseEnergy = 300000n;