+++ packages/common/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Changelog
 
-## 1.0.2 unreleased
+## 2.0.0
 
+### Added
+
+- Support deserializing version 2 schemas.
+- Support serializing parameters for contracts using version 2 schemas.
 - Support for deploying versioned smart contract modules. (This is done by not supplying the version field in the payload)
+
+### Breaking changes
+
+- `serializeInitContractParameters` and `serializeUpdateContractParameters` each have an additional parameter, which denotes the version of the schema provided. For existing users that are using V0 contracts, that parameter should be `SchemaVersion.V1`.
+- Deserialization of schemas have been changed: types and functions have been renamed and `getModuleFromBuffer` have an additional parameter.
 
 ## 1.0.1 2022-6-2